Refactoring: bulktree logic out of API call
